[IMP] fastapi: Always run tests at post-install
When loading the registry, Odoo takes rlock on the current thread to ensure the thead safety of the loading process. Nevertheless, dealing at the same time with lock on thread and the asyncio eventloop lead to the blockage of the eventloop. By running tests at post install, we avoid this kind of problem